Large migratory updates; see details.
This commit brings sobek-playground in line with the philosophy
discussed in the new Sobek 0.4.0 documentation. Essentially, this is an
area in which Sobek-centric code can be managed but does NOT have to
conform to the more strict coding and style regulations. Further, it
uses the existing Sobek assets and new integrated window module (again,
as of 0.4.0).